+++ ng_trajectory/selectors/fixed/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3.6
+"""Fixed selector.
+
+This is not a selector as other, but it just imitates
+one. Fixed selector takes points from the arguments
+and returns them upon calling.
+"""
+from .main import init, select