+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ## Unreleased
 ### Added
 - Selectors: Downsample factor for Curvature selector.
+- Selectors: Fixed selector that returns preset list of points.
 
 ## 1.4.3 - 2021-11-12
 ### Changed